Refactor - page abstractor
