+++ playwright-e2e/fixtures/create-fixture.ts
@@ -1,27 +1,52 @@
 import { test as base } from '@playwright/test';
+import { CommonComponents } from '../pages/CommonComponents';
 import { LoginPage } from '../pages/LoginPage';
 import { ManageCasePage } from '../pages/create-case/ManageCasePage';
-import { formAApplicationPage } from '../pages/formAApplicationPage';
-import { CommonComponents } from '../pages/commonComponents';
+import { SolicitorDetailsPage } from '../pages/create-case/SolicitorDetailsPage';
+import { DivorceDetailsPage } from '../pages/create-case/DivorceDetailsPage';
+import { ApplicantDetailsPage } from '../pages/create-case/ApplicantDetailsPage';
+import { RespondentRepresentedPage } from '../pages/create-case/RespondentRepresentedPage';
+import { NatureOfApplicationPage } from '../pages/create-case/NatureOfApplicationPage';
+import { PropertyAdjustmentPage } from '../pages/create-case/PropertyAdjustmentPage';
 
 type CreateFixtures = {
+  commonComponents: CommonComponents;
   loginPage: LoginPage;
   manageCasePage: ManageCasePage;
-  commonComponents: CommonComponents;
-  formAApplicationPage: formAApplicationPage;
+  solicitorDetailsPage: SolicitorDetailsPage;
+  divorceDetailsPage: DivorceDetailsPage;
+  applicantDetailsPage: ApplicantDetailsPage;
+  respondentRepresentedPage: RespondentRepresentedPage;
+  natureOfApplicationPage: NatureOfApplicationPage;
+  propertyAdjustmentPage: PropertyAdjustmentPage;
 };
 
 export const test = base.extend<CreateFixtures>({
+  commonComponents: async ({ page }, use) => {
+    await use(new CommonComponents(page));
+  },
   loginPage: async ({ page }, use) => {
     await use(new LoginPage(page));
   },
   manageCasePage: async ({ page }, use) => {
     await use(new ManageCasePage(page));
   },
-  commonComponents: async ({ page }, use) => {
-    await use(new CommonComponents(page));
+  solicitorDetailsPage: async ({ page }, use) => {
+    await use(new SolicitorDetailsPage(page));
   },
-  formAApplicationPage: async ({ page }, use) => {
-    await use(new formAApplicationPage(page));
+  divorceDetailsPage: async ({ page }, use) => {
+    await use(new DivorceDetailsPage(page));
   },
+  applicantDetailsPage: async ({ page }, use) => {
+    await use(new ApplicantDetailsPage(page));
+  },
+  respondentRepresentedPage: async ({ page }, use) => {
+    await use(new RespondentRepresentedPage(page));
+  },
+  natureOfApplicationPage: async ({ page }, use) => {
+    await use(new NatureOfApplicationPage(page));
+  },
+  propertyAdjustmentPage: async ({ page }, use) => {
+    await use(new PropertyAdjustmentPage(page));
+  }
 });
